Revised the algorithm.
Revised the algorithm and retested with dieharder and TestU01 for randomness.  eegl now runs as fast as the other aquila62 random number generators and as fast as the GSL version of taus2.   Increased the number of tests to match the tests in the other aquila62 repositories for random number generators.